[BE] user 정보 추가

+++ backend/src/main/java/com/capstone/samadhi/record/dto/TimeLineResponse.java
@@ -14,7 +14,7 @@ public record TimeLineResponse(
         String pose,
 
         @Schema(description = "해당 구간 점수", example = "95")
-        int score
+        float score
 ) {
 
     public static TimeLineResponse from(TimeLine timeLine) {
